requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
Deprecated==1.2.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-crispy-forms==1.9.0
django-icons==1.1.2
django-leaflet==0.26.0
django-map-widgets==0.3.0
django-postgresql==0.0.3
django-toolbelt==0.0.1
gitdb==4.0.2
GitPython==3.1.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
ogr==0.11.1
Pillow==7.0.0
postgis==1.0.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.47
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-gitlab==2.2.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
service-identity==18.1.0
six==1.14.0
smmap==3.0.1
sqlparse==0.3.1
static3==0.7.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
wrapt==1.12.1
zope.interface==5.0.2
